to do
Adding this markdown file to help track where I'm at in the project and
what is coming up. Eventually it could also track priorities and other
types of checklists. It's all to help stay organized, keep track of the
current status of the project, and know where to pick up next.